move fraud proof v0 to its own module
